# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Contract Lifecycle Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contracts/{id}/obligations:
    get:
      operationId: listObligations
      tags: [obligations]
      summary: List obligations for a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Obligations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Obligation"

    post:
      operationId: createObligation
      tags: [obligations]
      summary: Create an obligation for a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateObligationBody"
      responses:
        "201":
          description: Obligation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Obligation"
  /obligations/{id}/complete:
    post:
      operationId: completeObligation
      tags: [obligations]
      summary: Mark an obligation as completed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Obligation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Obligation"

  # Dashboard
...
components:
  schemas:
...
    Obligation:
      type: object
      properties:
        id:
          type: integer
        contractId:
          type: integer
        contractName:
          type: ["string", "null"]
        obligationType:
          type: string
        description:
          type: string
        dueDate:
          type: string
        status:
          type: string
          enum: [pending, completed, overdue, acknowledged]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - contractId
        - obligationType
        - description
        - dueDate
        - status
        - createdAt
    CreateObligationBody:
      type: object
      properties:
        obligationType:
          type: string
        description:
          type: string
        dueDate:
          type: string
        status:
          type: string
          enum: [pending, completed, overdue, acknowledged]
      required:
        - obligationType
        - description
        - dueDate
...